chore: share chopsticks instance between test runs
