+++ integration-tests/chopsticks/setupTests.ts
@@ -1,0 +1,6 @@
+import { startChopsticks, stopChopsticks } from "./src/chopsticks"
+
+export default async function setupTests() {
+  await startChopsticks()
+  return () => stopChopsticks()
+}
