+++ integration-tests/chopsticks/vitest.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from "vitest/config"
+import tsconfigPaths from "vite-tsconfig-paths"
+
+export default defineConfig({
+  test: {
+    include: ["**/*.spec.ts"],
+    disableConsoleIntercept: true,
+    globalSetup: ["./setupTests.ts"],
+  },
+  plugins: [tsconfigPaths()],
+})
